KML: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31614 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/kml/ogrkmllayer.cpp
@@ -442,7 +442,7 @@ OGRErr OGRKMLLayer::ICreateFeature( OGRFeature* poFeature )
             else
                 bHasWidth = TRUE;
             const char* pszColor = poPen->Color(bDefault);
-            int nColorLen = CPLStrnlen(pszColor, 10);
+            int nColorLen = static_cast<int>(CPLStrnlen(pszColor, 10));
             if( pszColor != NULL && pszColor[0] == '#' && !bDefault && nColorLen >= 7)
             {
                 char acColor[9] = {0};
